Update _getListCardByQueue in LearnApi.java

+++ app/src/main/java/com/born2go/lazzybee/db/api/LearnApi.java
@@ -123,7 +123,7 @@ public interface LearnApi {
      * @param queue
      *
      */
-    List<Card> _getListCardByQueue(long queue);
+    List<Card> _getListCardByQueue(int queue);
 
 
 }
